DOC Mark Fortran code as linguist-vendored
GitHub's linguist library sees the glmnet Fortran code and marks this repository as a Fortran project. It would be better to have it display as Python. Tell `linguist` that the Fortran code is vendored so that it will be ignored in summary statistics.

See https://github.com/github/linguist#vendored-code